# Improve multi-threading behaviour of indexer