@AHSauge 2025-08-09
I tested with more git repos and also limited the CPU cores, and concluded this is pretty tricky to tweak. I've landed on spawning one worker thread per CPU thread which seems to guarantee on-par performance or improvements across different repos and CPU core count.

That said, even with this patch, for some mysterious reason there are repos where there's no thread maxing out. Even artificially limiting the number of threads created to a situation where the diff grabbing thread really should max out, it still doesn't. This suggest some pretty severe performance issues in `git_diff_tree_to_tree`